fixed array filters and moved TestLayout
